requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
urllib3==2.2.0
pydantic==2.6.0